+++ esp-config/CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Crate prefixes and configuration keys are now separated by two underscores (`__`) (#2848)
+- Crate prefixes and configuration keys are now separated by `_CONFIG_` (#2848)
 
 ### Removed